Upgrade to DttSP 1.6.2, mostly. Fixed bugs in the Waterfall display. Other minor changes.

+++ trunk/Software/UWSDRGUI/DSPControl.cpp
@@ -27,7 +27,7 @@ const int RX_READER = 88;
 const int RINGBUFFER_SIZE = 100001;
 const int BLOCK_SIZE      = 2048;		// XXXX
 
-CDSPControl::CDSPControl(unsigned int sampleRate, float centreFreq) :
+CDSPControl::CDSPControl(float sampleRate, float centreFreq) :
 wxThread(),
 m_dttsp(NULL),
 m_sampleRate(sampleRate),
