+++ trunk/Software/UWSDRGUI/DataWriter.h
@@ -21,7 +21,7 @@
 
 class IDataWriter {
     public:
-	virtual bool open(unsigned int sampleRate, unsigned int blockSize) = 0;
+	virtual bool open(float sampleRate, unsigned int blockSize) = 0;
 	virtual void write(const float* buffer, unsigned int nSamples) = 0;
 	virtual void close() = 0;
 };
