+++ trunk/Software/UWSDRGUI/NullReader.h
@@ -31,14 +31,14 @@ class CNullReader : public wxThread, public IDataReader {
 
 	virtual void setCallback(IDataCallback* callback, int id);
 
-	virtual bool open(unsigned int sampleRate, unsigned int blockSize);
+	virtual bool open(float sampleRate, unsigned int blockSize);
 
 	virtual void* Entry();
 
 	virtual void close();
 
     private:
-	unsigned int   m_sampleRate;
+	float          m_sampleRate;
 	unsigned int   m_blockSize;
 	IDataCallback* m_callback;
 	int            m_id;
