+++ trunk/Software/UWSDRGUI/NullWriter.h
@@ -28,7 +28,7 @@ class CNullWriter : public IDataWriter {
 	CNullWriter();
 	virtual ~CNullWriter();
 
-	virtual bool open(unsigned int sampleRate, unsigned int blockSize);
+	virtual bool open(float sampleRate, unsigned int blockSize);
 	virtual void write(const float* buffer, unsigned int nSamples);
 	virtual void close();
