+++ trunk/Software/UWSDRGUI/SignalReader.h
@@ -31,7 +31,7 @@ class CSignalReader : public wxThread, public IDataReader {
 
 	virtual void setCallback(IDataCallback* callback, int id);
 
-	virtual bool open(unsigned int sampleRate, unsigned int blockSize);
+	virtual bool open(float sampleRate, unsigned int blockSize);
 
 	virtual void* Entry();
 
@@ -41,7 +41,7 @@ class CSignalReader : public wxThread, public IDataReader {
 	unsigned int   m_frequency;
 	float          m_noiseAmplitude;
 	float          m_signalAmplitude;
-	unsigned int   m_sampleRate;
+	float          m_sampleRate;
 	unsigned int   m_blockSize;
 	IDataCallback* m_callback;
 	int            m_id;
