+++ trunk/Software/UWSDRGUI/UWSDRData.h
@@ -34,7 +34,7 @@ class CUWSDRData : public wxEvtHandler, public IDataReader, public IDataWriter {
 
 	virtual void setCallback(IDataCallback* callback, int id);
 
-	virtual bool open(unsigned int sampleRate, unsigned int blockSize);
+	virtual bool open(float sampleRate, unsigned int blockSize);
 
 	virtual void write(const float* buffer, unsigned int nSamples);
